Support reconciliation of NamespaceTemplate changes
Detect changes in the NamespaceTemplate object by maintaining a hashMap
of checksum of the object. Compare the current version of the object with
checksum in the hashMap. If they are different, then provision the new changes
in all affected namespaces.

For now, implemented the backend to support Pod changes only.

Signed-off-by: Anusha Ragunathan <anusha.ragunathan@docker.com>